seperate match status and team score components from footbal score component
